# swagger.yaml
openapi: 3.0.3
info:
  title: API проекту
  version: 1.0.0
  description: Повна документація API для автентифікації, коментарів, проектів і задач.
servers:
  - url: http://localhost:5000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ResponseTaskDto:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
        description:
          type: string
        file:
          type: string
          nullable: true
        status:
          type: string
        comments:
          type: array
          nullable: true
          items:
            type: object
        answers:
          type: array
          nullable: true
          items:
            type: object
        assignedTasks:
          type: array
          nullable: true
          items:
            type: object
paths:
...
  /task/{taskId}/projects/{projectId}/users:
    post:
      summary: Додати користувача до задачі (треба бути creatер-ом)
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                  example: 5
      responses:
        '200':
          description: Повідомлення про додавання користувача
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Неавторизований
  /task/{taskId}/projects/{projectId}:
    patch:
      summary: Оновити задачу (треба бути creatер-ом)
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Оновлена задача
          content:
            application/json:
              schema:
                type: object
                properties:
                  task:
                    $ref: '#/components/schemas/ResponseTaskDto'
                  file:
                    type: string
                    nullable: true
        '401':
          description: Неавторизований

    delete:
      summary: Видалити задачу (треба бути creatер-ом)
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Повідомлення про видалення
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Задачу видалено"
        '401':
          description: Неавторизований
  /task/answers:
    post:
      summary: Створити відповідь до задачі
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                file:
                  type: string
                  format: binary
                userId:
                  type: string
                taskId:
                  type: string
      responses:
        '201':
          description: Створена відповідь
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer:
                    $ref: '#/components/schemas/ResponseAsnwersDto'
                  file:
                    type: string
                    nullable: true
        '401':
          description: Неавторизований
  /task/{taskId}/answers:
    patch:
      summary: Оновити відповідь до задачі
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                file:
                  type: string
                  format: binary
                userId:
                  type: string
                taskId:
                  type: string
      responses:
        '200':
          description: Оновлена відповідь
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer:
                    $ref: '#/components/schemas/ResponseAsnwersDto'
                  file:
                    type: string
                    nullable: true
        '401':
          description: Неавторизований